Moved the TopGeneratedDirectory and TopRawDirectoryToTexturePack enums to the asset pacakge
